Various changes
- Changed Vertrag mapping to properly enforce one-to-one constraint
between Vertrag and Immobilie
- Changed addMietvertrag() and addKaufvertrag() in ImmoService to throw
the ConstraintViolationException which gets thrown if you try to insert
a second contract against a house which already has one
- Changed VertragEditor to handle the ConstraintViolationException
thrown by the above methods
- Changed getAllWohnungenForMakler() and getAllHaeuserForMakler() in
ImmoService to take an additional boolean parameter which allows
restriction of the returned properties to only ones which do not already
have a contract assigned. Updated calls to these methods in ImmoService
and ImmobilienEdtitor to match the new signature.